[refactor]: Split pipeline tests into focused files by operation type
Split large pipeline_test.go file into smaller, more maintainable test files
organized by operation type to improve test organization and discoverability:

- pipeline_core_test.go: Core pipeline functionality and initialization
- pipeline_filter_test.go: Filter operation test coverage
- pipeline_sort_test.go: Sort operation test validation
- pipeline_limit_test.go: Limit operation test scenarios
- pipeline_advanced_test.go: Advanced features and optimization tests

All existing test coverage has been preserved while enhancing maintainability
and making tests easier to navigate and understand.
